blog_censurfridns_dk/requirements.txt
Django==1.8.2
django-bootstrap3==7.0.0
django-taggit==0.18.0
django-taggit-templatetags==0.2.5
django-modeltranslation==0.11